pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "oacs"
version = "0.1.0"
description = "Open Agent Context Standard proof-of-concept and Agent Context Shell"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "OACS Contributors" }]
dependencies = [
  "cryptography>=42",
  "fastapi>=0.110",
  "httpx>=0.27",
  "jsonschema>=4.21",
  "pydantic>=2.6",
  "rich>=13.7",
  "typer>=0.12",
  "uvicorn>=0.27",
]

[project.optional-dependencies]
crypto = ["argon2-cffi>=23.1"]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.5",
  "mypy>=1.10",
]
pq = ["liboqs-python>=0.10; platform_system != 'Windows'"]

[project.scripts]
acs = "oacs.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["oacs"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=oacs --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
packages = ["oacs"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
